+++ sample/Kmd.Logic.CitizenDocuments.Client.Sample/ConfigurationValidator.cs
@@ -17,8 +17,7 @@ namespace Kmd.Logic.CitizenDocuments.Client.Sample
             if (string.IsNullOrWhiteSpace(this._configuration.TokenProvider?.ClientId)
                 || string.IsNullOrWhiteSpace(this._configuration.TokenProvider?.ClientSecret)
                 || string.IsNullOrWhiteSpace(this._configuration.TokenProvider?.AuthorizationScope)
-                || !Guid.TryParse(this._configuration.SubscriptionId, out _)
-                || !Guid.TryParse(this._configuration.ConfigurationId, out _))
+                || !Guid.TryParse(this._configuration.SubscriptionId, out _))
             {
                 Log.Error(
                     "Invalid configuration. Please provide proper information to `appsettings.json`. Current data is: {@Settings}",