+++ CHANGES.md
@@ -1,3 +1,7 @@
+3.0.0
+* Changed the LogicTokenProviderFactory to ITokenProviderFactory
+* Exposed Create citizen Document Configuration in client 
+
 2.0.0
 * Make client assembly strongly named
 * Make client assembly versioned